Church sit animation code + made clone + fixed match again
added the sit animation in but the code doesnt play the animation very well so that still needs to be fixed later. Create a the clone for the church and made it into a prefab but it still needs to be duplicated several times. also for some reason the brighten match broke again but I managed to fix it.